Add files via upload
This study aims to analyze temperature change trends based on daily climate data obtained from BMKG. The time series method and linear regression model are used to identify the pattern of temperature change in 2024. Using Python and BMKG daily climate data, the data is processed and analyzed computationally. The results showed a downward trend in temperature in the last 6 months since the study began.